+++ docs/news.rst
@@ -6,6 +6,8 @@ Development (master)
 
 Features:
 
+  - Implement LoadTemplate.loadTemplate to load template
+    from .py[co], .py or .tmpl.
   - CheetahDirOwner caches compiled template in the template directory.
   - CheetahDirOwner now silently ignores errors on compiled templates
     writing. To get tracebacks set CheetahDirOwner.debuglevel = 1.